Unit 6b challenge work
The power up now finally gets collected, it applies a speed boost temporarily and they spawn randomly. I also made two variations which have varying speed and modifiers applied when collected which are random chosen between when spawning. (in order to remove the speed modifier after a duration of time I had to make them invisible and have no collision for the duration because deleting them early stops it from being able to undo the speed up. The back boundary was therefore increased to give ample time for the powerups when activated to wear off. I also could have implemented something that stops the move forward script for them but it was easier to  just increase the back boundary.) Also upon coming in contact with a powerup the player emits particles to bring attention to the collection of the powerup.